+++ src/stockpyl/supply_chain_network.py
@@ -7,7 +7,7 @@
 # ===============================================================================
 
 """
-.. include:: ../globals.inc
+.. include:: ../../globals.inc
 
 This module contains the ``SupplyChainNetwork`` class.