+++ src/stockpyl/supply_chain_node.py
@@ -7,7 +7,7 @@
 # ===============================================================================
 
 """
-.. include:: ../globals.inc
+.. include:: ../../globals.inc
 
 This module contains the ``SupplyChainNode`` class, which is a stage or node
 in a supply chain network.